requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.9.0
insightface>=0.7.3
pymongo>=4.0.0
requests>=2.26.0
scipy>=1.7.0
flask>=2.0.0
werkzeug>=2.0.0
onnxruntime>=1.15.0
onnx>=1.14.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
gunicorn>=20.1.0
gevent>=21.12.0
